+++ src/scancode_config.py
@@ -77,7 +77,7 @@ def _create_dir(location):
 
 # in case package is not installed or we do not have setutools/pkg_resources
 # on hand fall back to this version
-__version__ = '31.2.1'
+__version__ = '31.2.3'
 
 # used to warn user when the version is out of date
 __release_date__ = datetime.datetime(2022, 10, 5)
@@ -87,7 +87,7 @@ __release_date__ = datetime.datetime(2022, 10, 5)
 __output_format_version__ = '2.0.0'
 
 #
-spdx_license_list_version = '3.17'
+spdx_license_list_version = '3.18'
 
 try:
     from pkg_resources import get_distribution, DistributionNotFound